Add death and basic inventory.

+++ src/BloomEffect.cpp
@@ -64,8 +64,8 @@ void BloomEffect::blurMultipass(RenderTextureArray& renderTextures)
 
 	for (std::size_t count = 0; count < 2; ++count)
 	{
-		blur(renderTextures[0], renderTextures[1], sf::Vector2f(0.f, 1.f / textureSize.y));
-		blur(renderTextures[1], renderTextures[0], sf::Vector2f(1.f / textureSize.x, 0.f));
+		blur(renderTextures[0], renderTextures[1], sf::Vector2f(0.f, 1.f / static_cast<float>(textureSize.y)));
+		blur(renderTextures[1], renderTextures[0], sf::Vector2f(1.f / static_cast<float>(textureSize.x), 0.f));
 	}
 }
